+++ src/compas_cem/equilibrium/state.py
@@ -1,0 +1,5 @@
+from typing import NamedTuple
+
+
+class EquilibriumState(NamedTuple):
+    pass