+++ CHANGELOG.rst
@@ -11,6 +11,11 @@ Unreleased
 
 **Added**
 
+- Added `PolylineConstraint` component to grasshopper plugin.
+- Implemented `constraints.PolylineConstraint`.
+- Added `EdgeDirectionConstraint` component to grasshopper plugin.
+- Implemented `constraints.EdgeDirectionConstraint`.
+
 **Changed**
 
 **Fixed**